+++ src/main/java/duke/exception/DukeInvalidFileException.java
@@ -1,0 +1,10 @@
+package duke.exception;
+
+/**
+ * Exceptions that arise when the file is corrupted.
+ */
+public class DukeInvalidFileException extends DukeRuntimeException {
+    public DukeInvalidFileException(String message) {
+        super(message);
+    }
+}
