+++ src/main/java/duke/task/TaskStorage.java
@@ -5,7 +5,9 @@ import java.io.FileNotFoundException;
 import java.util.ArrayList;
 import java.util.Scanner;
 
-import duke.Duke;
+import duke.exception.DukeInvalidCommandException;
+import duke.exception.DukeInvalidFileException;
+import duke.exception.DukeWrongCommandFormatException;
 
 /**
  * Duke.TaskStorage class is used to store all the tasks in the list. It also has the capability of saving the tasks to
@@ -27,7 +29,7 @@ public class TaskStorage {
         this.file = new File(filePath);
         try {
             this.loadFromFile();
-        } catch (FileNotFoundException | Duke.DukeWrongCommandFormatException | Duke.DukeInvalidFileException e) {
+        } catch (FileNotFoundException | DukeWrongCommandFormatException | DukeInvalidFileException e) {
             try {
                 this.file.getParentFile().mkdirs();
                 this.file.createNewFile();
@@ -46,7 +48,7 @@ public class TaskStorage {
         this.file = new File(newFilePath);
         try {
             this.loadFromFile();
-        } catch (FileNotFoundException | Duke.DukeWrongCommandFormatException | Duke.DukeInvalidFileException e) {
+        } catch (FileNotFoundException | DukeWrongCommandFormatException | DukeInvalidFileException e) {
             try {
                 this.file.getParentFile().mkdirs();
                 this.file.createNewFile();
@@ -66,7 +68,7 @@ public class TaskStorage {
         try {
             assert(input != null);
             task = Task.createTask(input);
-        } catch (Duke.DukeInvalidCommandException | Duke.DukeWrongCommandFormatException e) {
+        } catch (DukeInvalidCommandException | DukeWrongCommandFormatException e) {
             return e.getMessage();
         }
 
@@ -161,7 +163,7 @@ public class TaskStorage {
     }
 
     private void loadFromFile()
-            throws FileNotFoundException, Duke.DukeWrongCommandFormatException, Duke.DukeInvalidFileException {
+            throws FileNotFoundException, DukeWrongCommandFormatException, DukeInvalidFileException {
         Scanner sc = new Scanner(file);
         while (sc.hasNextLine()) {
             String fileTask = sc.nextLine();
